Move these function to where they belong
See https://github.com/instedd/surveda/pull/1901#discussion_r656349698
